feat: Add getCategories function to fetch and return unique categories
This commit adds a new function, `getCategories`, to the `categoryActions.ts` file. The function uses the `prisma` library to fetch categories from the database and returns an array of unique categories. If there are no categories, it returns an error message. This function will be used to populate the category dropdown in the UI.