feat: Update next.config.js and tailwind.config.ts
This commit updates the `next.config.js` file to change the hostname in one of the image sources from "tailus.io" to "images.unsplash.com". This change ensures that the images are loaded from the correct source.

Additionally, the `tailwind.config.ts` file is modified to include a new keyframe animation called "scroll". This animation is used to create a scrolling effect on certain elements of the website.

These changes improve the functionality and visual appeal of the website.